% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrtcrit4.R
\name{lrtcrit4}
\alias{lrtcrit4}
\title{Find test statistic value from a given data and corresponding critical value}
\usage{
lrtcrit4(X1, X2, X3, X4, alpha)
}
\arguments{
\item{X1}{a real vector}

\item{X2}{a real vector}

\item{X3}{a real vector}

\item{X4}{a real vector}

\item{alpha}{real number between 0 and 1 called significance level}
}
\value{
a numeric vector
}
\description{
More detailed description
}
\examples{
X1=c(4.28,2.56,0.83,1.41,2.18,1.47,2.66,-0.73,3.58,1.55,4.2,2.98,1.12,0.22,4.3)
X2=c(2.12,-1.03,3.37,6.95,2.42,4.22,5.97,0.84,4.57,-0.16,2.18,2.56)
X3=c(4.76,3.85,3.54,3.55,2.73,2.61,3.81,3.61,3.28,3.01,3.67,1.56,4.6,4.74)
X4=c(4.7,3.74,5.82,2.03,4.9,4.69,1.45,4.9,2.5,3.33)
lrtcrit4(X1,X2,X3,X4,0.05)

}
